refs(release_health): Move releasemonitor into the `release_health` module.
Refactoring this so that it's located with other `release_health` related code. Since we're porting
this over to metrics, we'll use backends to allow us to switch and compare between sessions/metrics.
